즐겨찾기 ajax로 재수정 myhome 사진 뿌려주는거 6개로 변경

+++ TeamProject0503/src/net/rating/db/RatingDAO.java
@@ -288,9 +288,8 @@ public class RatingDAO {
 				pstmt = con.prepareStatement(sql);
 				pstmt.setInt(1, mv_num);
 				rs = pstmt.executeQuery();
-				
 				if(rs.next()){
-					avg=rs.getInt("avg");
+					avg=rs.getFloat("avg");
 				}
 				
 			} catch(Exception e) {
